fixed issue with decoded entity modes and multiple line strings and also completed unit tests
