--- test/decode-test.js
@@ -1,10 +1,0 @@
-'use strict';
-
-// Load external dependencies.
-
-// Load internal dependencies.
-
-// Run test suite.
-describe('morjs.decode', function() {
-  // TODO: Complete
-});
