+++ test/fixtures/characters.txt
@@ -53,32 +53,32 @@ _
 $
 @
 Ä
-Ä
-Ä
+Æ
+Ą
 À
-À
+Å
 Ç
-Ç
-Ç
+Ĉ
+Ć
 Š
 Ð
 Ś
 È
-È
+Ł
 É
-É
-É
+Đ
+Ę
 Ĝ
-Š
+Ĥ
 Ĵ
 Ź
 Ñ
-Ñ
+Ń
 Ö
-Ö
-Ö
+Ø
+Ó
 Ŝ
 Þ
 Ü
-Ü
+Ŭ
 Ż
