+++ test/fixtures/encoded.txt
@@ -24,32 +24,6 @@
 -··-
 -·--
 --··
-·-
--···
--·-·
--··
-·
-··-·
---·
-····
-··
-·---
--·-
-·-··
---
--·
----
-·--·
---·-
-·-·
-···
--
-··-
-···-
-·--
--··-
--·--
---··
 -----
 ·----
 ··---
@@ -78,36 +52,6 @@
 ·-··-·
 ···-··-
 ·--·-·
-·-·-
-·-·-
-·-·-
-·--·-
-·--·-
--·-··
--·-··
--·-··
-----
-··--·
-···-···
-·-··-
-·-··-
-··-··
-··-··
-··-··
---·-·
-----
-·---·
---··-·
---·--
---·--
----·
----·
----·
-···-·
-·--··
-··--
-··--
---··-
 ·-·-
 ·-·-
 ·-·-
